fixed one bug with active user

+++ controller/task.php
@@ -57,7 +57,7 @@
         $returned_useractive = $row['useractive'];
         $returned_loginattempts = $row['loginattempts'];
 
-        if ($returned_useractiv !== 'Y') {
+        if ($returned_useractive !== 'Y') {
             $response = new Response();
             $response->setHttpStatusCode(401);
             $response->setSuccess(false);
@@ -91,7 +91,6 @@
         $response->addMessage('There was an issue authenticating - please try again');
         $response->send();
         exit();
-        // finished at 16:00 for 030Add Authentication video
     }
 
     // end authentication script
@@ -110,8 +109,9 @@
 
         if($_SERVER['REQUEST_METHOD'] === 'GET') {
             try{
-                $query = $readDB->prepare('select id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed from tbltasks where id = :taskid');
+                $query = $readDB->prepare('select id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed from tbltasks where id = :taskid and userid = :userid');
                 $query->bindParam(':taskid', $taskid, PDO::PARAM_INT);
+                $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
                 $query->execute();
 
                 $rowCount = $query->rowCount();
@@ -160,8 +160,9 @@
             }
         } elseif ($_SERVER['REQUEST_METHOD'] === 'DELETE') {
             try{
-                $query = $writeDB->prepare('delete from tbltasks where id = :taskid');
+                $query = $writeDB->prepare('delete from tbltasks where id = :taskid and userid = :userid');
                 $query->bindParam(':taskid', $taskid, PDO::PARAM_INT);
+                $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
                 $query->execute();
 
                 $rowCount = $query->rowCount();
@@ -251,8 +252,9 @@
                     exit;
                 }
 
-                $query = $writeDB->prepare('select id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed from tbltasks where id = :taskid');
+                $query = $writeDB->prepare('select id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed from tbltasks where id = :taskid and userid = :userid');
                 $query->bindParam(':taskid', $taskid, PDO::PARAM_INT);
+                $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
                 $query->execute();
 
                 $rowCount = $query->rowCount();
@@ -270,7 +272,7 @@
                     $task = new Task($row['id'], $row['title'], $row['description'], $row['deadline'], $row['completed']);
                 }
 
-                $queryString = "update tbltasks set ".$queryFields." where id = :taskid";
+                $queryString = "update tbltasks set ".$queryFields." where id = :taskid and userid = :userid";
                 $query = $writeDB->prepare($queryString);
 
                 if($title_updated === true){
@@ -298,6 +300,7 @@
                 }
 
                 $query->bindParam(':taskid', $taskid, PDO::PARAM_INT);
+                $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
                 $query->execute();
 
                 $rowCount = $query->rowCount();
@@ -311,8 +314,9 @@
                     exit;
                 }
 
-                $query = $writeDB->prepare('select id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed from tbltasks where id = :taskid');
+                $query = $writeDB->prepare('select id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed from tbltasks where id = :taskid and userid = :userid');
                 $query->bindParam(':taskid', $taskid, PDO::PARAM_INT);
+                $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
                 $query->execute();
 
                 $rowCount = $query->rowCount();
@@ -384,8 +388,9 @@
 
         if($_SERVER['REQUEST_METHOD'] === 'GET') {
             try{
-                $query = $readDB->prepare('select id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed from tbltasks where completed = :completed');
+                $query = $readDB->prepare('select id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed from tbltasks where completed = :completed and userid = :userid');
                 $query->bindParam(':completed', $completed, PDO::PARAM_STR);
+                $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
                 $query->execute();
 
                 $rowCount = $query->rowCount();
@@ -449,7 +454,8 @@
             $limitPerPage = 20;
 
             try{
-                $query = $readDB->prepare('select count(id) as totalNoOfTasks from tbltasks');
+                $query = $readDB->prepare('select count(id) as totalNoOfTasks from tbltasks where userid = :userid');
+                $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
                 $query->execute();
 
                 $row = $query->fetch(PDO::FETCH_ASSOC);
@@ -473,7 +479,8 @@
 
                 $offset = ($page == 1 ? 0 : ($limitPerPage*($page-1)));
 
-                $query = $readDB->prepare('select id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed from tbltasks limit :pglimit offset :offset');
+                $query = $readDB->prepare('select id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed from tbltasks where userid = :userid limit :pglimit offset :offset');
+                $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
                 $query->bindParam(':pglimit', $limitPerPage, PDO::PARAM_INT);
                 $query->bindParam(':offset', $offset, PDO::PARAM_INT);
                 $query->execute();
@@ -530,7 +537,8 @@
     } elseif (empty($_GET)) {
         if($_SERVER['REQUEST_METHOD'] === 'GET') {
             try{
-                $query = $readDB->prepare('select id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed from tbltasks');
+                $query = $readDB->prepare('select id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed from tbltasks where userid = :userid');
+                $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
                 $query->execute();
 
                 $rowCount = $query->rowCount();
@@ -614,11 +622,12 @@
                 $deadline = $newTask->getDeadline();
                 $completed = $newTask->getCompleted();
 
-                $query = $writeDB->prepare('insert into tbltasks (title, description, deadline, completed) values (:title, :description, STR_TO_DATE(:deadline, \'%d/%m/%Y %H:%i\'), :completed)');
+                $query = $writeDB->prepare('insert into tbltasks (title, description, deadline, completed, userid) values (:title, :description, STR_TO_DATE(:deadline, \'%d/%m/%Y %H:%i\'), :completed, :userid)');
                 $query->bindParam(':title', $title, PDO::PARAM_STR);
                 $query->bindParam(':description', $description, PDO::PARAM_STR);
                 $query->bindParam(':deadline', $deadline, PDO::PARAM_STR);
                 $query->bindParam(':completed', $completed, PDO::PARAM_STR);
+                $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
                 $query->execute();
 
                 $rowCount = $query->rowCount();
@@ -634,8 +643,9 @@
                 
                 $lastTaskID = $writeDB->lastInsertId();
 
-                $query = $writeDB->prepare('select id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed from tbltasks where id = :taskid');
+                $query = $writeDB->prepare('select id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed from tbltasks where id = :taskid and userid = :userid');
                 $query->bindParam(':taskid', $lastTaskID, PDO::PARAM_INT);
+                $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
                 $query->execute();
 
                 $rowCount = $query->rowCount();
@@ -705,6 +715,6 @@
         $response->send();
         exit;
     }
-
+// finished here at 30:00 at video 030 Add authentication
 
 ?>
